requirements.txt
faster-whisper>=0.10.0
pyaudio>=0.2.13
pyttsx3>=2.90
PyYAML>=6.0
pyautogui>=0.9.54
numpy>=1.23.0
keyboard>=0.13.5
Pillow>=9.0.0
